State Share of National GST Collection

Each state's percentage share of national total for a given period.

gst_state_share(
  component = "total",
  financial_year = NULL,
  from_month = NULL,
  to_month = NULL,
  exclude_other = TRUE
)

Arguments

component

character(1). Default "total".

financial_year

character vector. NULL = all.

from_month

character. Start month.

to_month

character. End month.

exclude_other

logical. Default TRUE.

Value

A data.table with state, region, the component total, and share_pct.

Examples

gst_state_share(financial_year = "2023-24")[1:10]
#>             state      total  region share_pct
#>            <char>      <num>  <char>     <num>
#>  1:               4036498.40   Other 72.669140
#>  2:   Maharashtra  320116.84    West  5.763068
#>  3:     Karnataka  145265.89   South  2.615224
#>  4:       Gujarat  125168.37    West  2.253408
#>  5:    Tamil Nadu  121329.11   South  2.184290
#>  6:       Haryana  102914.46   North  1.852771
#>  7: Uttar Pradesh  101692.58 Central  1.830773
#>  8:         Delhi   66445.39   North  1.196217
#>  9:   West Bengal   62613.00    East  1.127223
#> 10:     Telangana   59942.31   South  1.079142